+++ include/wlr/types/wlr_output.h
@@ -84,7 +84,7 @@ void wlr_output_effective_resolution(struct wlr_output *output,
 void wlr_output_make_current(struct wlr_output *output);
 void wlr_output_swap_buffers(struct wlr_output *output);
 void wlr_output_set_gamma(struct wlr_output *output,
-	uint16_t size, uint16_t *r, uint16_t *g, uint16_t *b);
-uint16_t wlr_output_get_gamma_size(struct wlr_output *output);
+	uint32_t size, uint16_t *r, uint16_t *g, uint16_t *b);
+uint32_t wlr_output_get_gamma_size(struct wlr_output *output);
 
 #endif
